requirements.txt
fastapi
uvicorn
python-multipart
pillow
python-dotenv
httpx
openai